CMake and code Cleanup (#60)
* Cleanup CMake,  remove unknown pragma warnings
* Remove NDI example, some general code cleanup

+++ src/apps/sound/main.cpp
@@ -8,18 +8,17 @@
 
 #include "sgct.h"
 
-//include open AL
 #ifdef __APPLE__
-    #include <OpenAL/al.h>
-    //#include <ALUT/alut.h> //if using ALUT framework
+#include <OpenAL/al.h>
+// #include <ALUT/alut.h>
 #else
-    #include <AL/al.h>
+#include <AL/al.h>
 #endif
 #include <AL/alut.h>
 
-sgct::Engine * gEngine;
+sgct::Engine* gEngine;
 
-//callbacks
+// callbacks
 void myInitOGLFun();
 void myDrawFun();
 void myPreSyncFun();
@@ -28,40 +27,38 @@ void myEncodeFun();
 void myDecodeFun();
 void myCleanUpFun();
 
-//other functions
-void setAudioSource(ALuint &buffer,ALuint &source, const char * filename);
+// other functions
+void setAudioSource(ALuint& buffer, ALuint& source, const char* filename);
 
-//open AL data
+// open AL data
 ALuint audio_buffer0 = AL_NONE;
 ALuint source0;
 glm::vec4 audioPos;
 
-//pther variables
+// other variables
 sgct::SharedDouble curr_time(0.0);
-float speed = 25.0f;
+float speed = 25.f;
 float radius = 7.4f;
-float objectRadius = 5.0f;
+float objectRadius = 5.f;
 float PI = 3.141592654f;
-float domeTilt = 27.0f;
+float domeTilt = 27.f;
 
-sgct_utils::SGCTSphere *sphere;
-sgct_utils::SGCTDome * dome;
+sgct_utils::SGCTSphere* sphere;
+sgct_utils::SGCTDome* dome;
 
-int main( int argc, char* argv[] )
-{
-    gEngine = new sgct::Engine( argc, argv );
+int main(int argc, char* argv[]) {
+    gEngine = new sgct::Engine(argc, argv);
 
     // Bind your functions
-    gEngine->setInitOGLFunction( myInitOGLFun );
-    gEngine->setDrawFunction( myDrawFun );
-    gEngine->setPreSyncFunction( myPreSyncFun );
-    gEngine->setPostSyncPreDrawFunction( myPostSyncPreDrawFun );
-    gEngine->setCleanUpFunction( myCleanUpFun );
+    gEngine->setInitOGLFunction(myInitOGLFun);
+    gEngine->setDrawFunction(myDrawFun);
+    gEngine->setPreSyncFunction(myPreSyncFun);
+    gEngine->setPostSyncPreDrawFunction(myPostSyncPreDrawFun);
+    gEngine->setCleanUpFunction(myCleanUpFun);
     sgct::SharedData::instance()->setEncodeFunction(myEncodeFun);
     sgct::SharedData::instance()->setDecodeFunction(myDecodeFun);
 
-    if( !gEngine->init() )
-    {
+    if (!gEngine->init()) {
         delete gEngine;
         return EXIT_FAILURE;
     }
@@ -73,19 +70,18 @@ int main( int argc, char* argv[] )
     delete gEngine;
 
     // Exit program
-    exit( EXIT_SUCCESS );
+    exit(EXIT_SUCCESS);
 }
 
-void myInitOGLFun()
-{
-    glEnable( GL_DEPTH_TEST );
+void myInitOGLFun() {
+    glEnable(GL_DEPTH_TEST);
 
     sphere = new sgct_utils::SGCTSphere(0.5f, 8);
-    dome = new sgct_utils::SGCTDome(7.4f, 165.0f, 36, 10);
+    dome = new sgct_utils::SGCTDome(7.4f, 165.f, 36, 10);
 
     alutInit(NULL, 0);
 
-    //Check for errors if any
+    // Check for errors if any
     sgct::MessageHandler::printInfo("ALUT init: %s", alutGetErrorString(alutGetError()));
 
     setAudioSource(audio_buffer0, source0, "file1.wav");
@@ -94,76 +90,69 @@ void myInitOGLFun()
     alListener3f(AL_POSITION, userPos.x, userPos.y, userPos.z);
     alListener3f(AL_VELOCITY, 0.0f, 0.0f, 0.0f);
 
-    alSourcef(source0, AL_PITCH, 1.0f);
-    alSourcef(source0, AL_GAIN, 1.0f);
-    alSource3f(source0, AL_POSITION, 0.0f, 0.0f, 0.0f);
-    alSource3f(source0, AL_VELOCITY, 0.0f, 0.0f, 0.0f);
+    alSourcef(source0, AL_PITCH, 1.f);
+    alSourcef(source0, AL_GAIN, 1.f);
+    alSource3f(source0, AL_POSITION, 0.f, 0.f, 0.f);
+    alSource3f(source0, AL_VELOCITY, 0.f, 0.f, 0.f);
     alSourcei(source0, AL_LOOPING, AL_TRUE);
 
     alSourcePlay(source0);
 }
 
-void myPostSyncPreDrawFun()
-{
-    float angle = glm::radians( static_cast<float>( curr_time.getVal() ) * speed );
+void myPostSyncPreDrawFun() {
+    float angle = glm::radians(static_cast<float>( curr_time.getVal()) * speed);
     glm::vec4 p;
-    p.x = objectRadius * sinf(angle);
-    p.z = objectRadius * cosf(angle);
-    p.y = 2.0f;
+    p.x = objectRadius * std::sin(angle);
+    p.z = objectRadius * std::cos(angle);
+    p.y = 2.f;
 
-    glm::mat4 rotMat = glm::mat4(1.0f);
-    rotMat = glm::rotate( rotMat, -domeTilt, glm::vec3(1.0f, 0.0f, 0.0f ) );
+    glm::mat4 rotMat = glm::mat4(1.f);
+    rotMat = glm::rotate(rotMat, -domeTilt, glm::vec3(1.f, 0.f, 0.f));
     audioPos = rotMat * p;
 
     alSource3f(source0, AL_POSITION, audioPos.x, audioPos.y, audioPos.z);
 }
 
-void myDrawFun()
-{
-    glLineWidth(2.0f);
+void myDrawFun() {
+    glLineWidth(2.f);
     glEnable(GL_BLEND);
     glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 
     glPushMatrix();
-        glTranslatef(audioPos.x, audioPos.y, audioPos.z);
-        glColor4f(1.0f, 0.4f, 0.1f, 0.8f);
-        //wireframe mode
-        glPolygonMode( GL_FRONT_AND_BACK, GL_LINE );
-        sphere->draw();
-        //reset to normal rendering
-        glPolygonMode( GL_FRONT_AND_BACK, GL_FILL );
+    glTranslatef(audioPos.x, audioPos.y, audioPos.z);
+    glColor4f(1.f, 0.4f, 0.1f, 0.8f);
+    //wireframe mode
+    glPolygonMode(GL_FRONT_AND_BACK, GL_LINE);
+    sphere->draw();
+    //reset to normal rendering
+    glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
     glPopMatrix();
 
     glPushMatrix();
-        glColor4f(0.0f, 0.4f, 1.0f, 0.8f);
-        dome->draw();
+    glColor4f(0.f, 0.4f, 1.f, 0.8f);
+    dome->draw();
     glPopMatrix();
 
     glDisable(GL_BLEND);
 }
 
-void myPreSyncFun()
-{
+void myPreSyncFun() {
     //set the time only on the master
-    if( gEngine->isMaster() )
-    {
+    if (gEngine->isMaster()) {
         //get the time in seconds
-        curr_time.setVal( sgct::Engine::getTime() );
+        curr_time.setVal(sgct::Engine::getTime());
     }
 }
 
-void myEncodeFun()
-{
-    sgct::SharedData::instance()->writeDouble( &curr_time );
+void myEncodeFun() {
+    sgct::SharedData::instance()->writeDouble(&curr_time);
 }
 
-void myDecodeFun()
-{
-    sgct::SharedData::instance()->readDouble( &curr_time );
+void myDecodeFun() {
+    sgct::SharedData::instance()->readDouble(&curr_time);
 }
 
-void myCleanUpFun()
-{
+void myCleanUpFun() {
     alDeleteSources(1, &source0);
     alDeleteBuffers(1, &audio_buffer0);
     alutExit();
@@ -172,15 +161,16 @@ void myCleanUpFun()
     delete dome;
 }
 
-void setAudioSource(ALuint &buffer, ALuint &source, const char * filename)
-{
+void setAudioSource(ALuint& buffer, ALuint& source, const char* filename) {
     alGenBuffers(1, &buffer);
     alGenSources(1, &source);
 
     buffer = alutCreateBufferFromFile(filename);
-    if( buffer == AL_NONE )
-    {
-        sgct::MessageHandler::printInfo("Failed to read audio file '%s', error: %s", filename, alutGetErrorString( alutGetError() ));
+    if (buffer == AL_NONE) {
+        sgct::MessageHandler::printInfo(
+            "Failed to read audio file '%s', error: %s",
+            filename, alutGetErrorString(alutGetError())
+        );
     }
     alSourcei(source, AL_BUFFER, buffer);
 }
